Fix size_t -> uint32_t on a 64 bits machine.
git-svn-id: https://vdsf.svn.sourceforge.net/svnroot/vdsf/trunk@890 53c26a9f-4021-0410-9f73-c4284c400f72

+++ src/API/Tests/HashMap/DeletePass.c
@@ -31,7 +31,7 @@ int main( int argc, char * argv[] )
    int errcode;
    const char * key  = "My Key";
    const char * data = "My Data";
-   size_t length, keyLength;
+   uint32_t length, keyLength;
    char buffer[20], keyBuff[20];
    psoObjectDefinition mapDef = { 
       PSO_HASH_MAP,
