+++ src/API/Tests/HashMap/InsertPass.c
@@ -32,7 +32,7 @@ int main( int argc, char * argv[] )
    const char * key  = "My Key";
    const char * data = "My Data";
    char buffer[20], keyBuff[20];
-   size_t length, keyLength;
+   uint32_t length, keyLength;
    psoObjectDefinition mapDef = { 
       PSO_HASH_MAP, 
       1,
