+++ src/C++/Tests/FastMap/Delete.cpp
@@ -35,7 +35,7 @@ int main( int argc, char * argv[] )
 
    const char * key  = "My Key";
    const char * data = "My Data";
-   size_t length, keyLength;
+   uint32_t length, keyLength;
    char buffer[20], keyBuff[20];
    int rc;
    psoObjectDefinition folderDef;
